+++ package_monitor/__init__.py
@@ -1,4 +1,4 @@
 default_app_config = "package_monitor.apps.PackageMonitorConfig"
 
-__version__ = "1.8.1"
+__version__ = "1.9.0"
 __title__ = "Package Monitor"