+++ CHANGELOG.md
@@ -7,11 +7,23 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
-### Added
+## [1.9.0] - 2023-04-22
 
 ### Changed
 
+- Added localization to enable translations
+
+### Changed
+
+- Now shows link to PyPI project page instead of the project's homepage
+- Build process migrated to PEP 621
+- Added support for AA 4 / Django 4
+- Dropped support for AA 3 / Django 3.1
+
 ### Fixed
+
+- PEP 660 packages are not detected as editable
+- Does not show homepage for some packages (#6)
 
 ## [1.8.1] - 2023-01-04